+++ packages/colony-js-client/src/TokenClient/index.js
@@ -48,7 +48,7 @@ export default class TokenClient extends ContractClient {
   */
   approve: TokenClient.Sender<
     {
-      user: Address, // The address that will be approved for the allowance (the token `spender`).
+      address: Address, // The address that will be approved for the allowance (the token `spender`).
       amount: BigNumber, // The amount of tokens that will be approved (the amount `allowed`).
     },
     {
@@ -64,11 +64,11 @@ export default class TokenClient extends ContractClient {
   >;
 
   /*
-  Burn tokens. This is a `DSToken` function that can only be called by the token `owner`. When a colony contract address is assigned as the token `owner`, this function can only be called by the user assigned the `FOUNDER` authority role.
+  Burn tokens. This function can only be called by the token owner or an address with authority.
   */
   burn: TokenClient.Sender<
     {
-      user: Address, // The address from which the tokens will be burned.
+      address: Address, // The address from which the tokens will be burned.
       amount: BigNumber, // The amount of tokens that will be burned.
     },
     {
@@ -108,7 +108,7 @@ export default class TokenClient extends ContractClient {
   getAllowance: TokenClient.Caller<
     {
       sourceAddress: Address, // The address that approved the allowance (the token `owner`).
-      user: Address, // The address that was approved for the allowance (the token `spender`).
+      address: Address, // The address that was approved for the allowance (the token `spender`).
     },
     {
       amount: BigNumber, // The amount of tokens that were approved (the amount `allowed`).
@@ -199,11 +199,11 @@ export default class TokenClient extends ContractClient {
   >;
 
   /*
-  Mint new tokens. This is a `DSToken` function that can only be called by the token `owner`. When a colony contract address is assigned as the token `owner`, this function can only be called by the user assigned the `FOUNDER` authority role.
+  Mint new tokens. This function can only be called by the token owner or an address with authority.
   */
   mint: TokenClient.Sender<
     {
-      user: Address, // The address that will receive the minted tokens.
+      address: Address, // The address that will receive the minted tokens.
       amount: BigNumber, // The amount of tokens that will be minted.
     },
     {
@@ -219,11 +219,11 @@ export default class TokenClient extends ContractClient {
   >;
 
   /*
-  Assign an account the `ADMIN` authority role within a colony.
+  Assign an account the token authority role within a colony.
   */
   setAuthority: TokenClient.Sender<
     {
-      authority: Address, // The address that will be assigned the `ADMIN` authority role.
+      authority: Address, // The address that will be assigned the token authority role.
     },
     {
       LogSetAuthority: LogSetAuthority,
@@ -320,7 +320,7 @@ export default class TokenClient extends ContractClient {
     const amount = ['amount', 'bigNumber'];
     const sourceAddress = ['sourceAddress', 'address'];
     const destinationAddress = ['destinationAddress', 'address'];
-    const user = ['user', 'address'];
+    const address = ['address', 'address'];
 
     // Events
     this.addEvent('Transfer', [
@@ -351,7 +351,7 @@ export default class TokenClient extends ContractClient {
     });
     this.addCaller('getAllowance', {
       functionName: 'allowance',
-      input: [sourceAddress, user],
+      input: [sourceAddress, address],
       output: [amount],
     });
     this.addCaller('isLocked', {
@@ -368,13 +368,13 @@ export default class TokenClient extends ContractClient {
       input: [sourceAddress, destinationAddress, amount],
     });
     this.addSender('approve', {
-      input: [user, amount],
+      input: [address, amount],
     });
     this.addSender('mint', {
-      input: [user, amount],
+      input: [address, amount],
     });
     this.addSender('burn', {
-      input: [user, amount],
+      input: [address, amount],
     });
     this.addSender('setOwner', {
       input: [['owner', 'address']],